requirements.txt
python-telegram-bot[job-queue]
python-telegram-bot[webhooks]
requests
aiohttp
python-dotenv